implement new gamma correction

+++ Direct2dViewer.h
@@ -94,8 +94,8 @@ public:
 	HWND getWindowHandler();
 	void SetGammaValue
 	(
-		UINT white,
-		UINT black
+		UINT16 white,
+		UINT16 black
 	);
 
 private:
@@ -133,7 +133,7 @@ private:
 	ID2D1Bitmap *m_pBitmap;
 	ID2D1Effect *linearTransferEffect;
 	FLOAT _gamma_amplitude = 1;
-	FLOAT _gamma_offset = 0;
+	UINT16 _gamma_offset = 0;
 
 	struct _BitmapSource
 	{
