+++ Direct2dViewer_c.cpp
@@ -32,7 +32,7 @@ HWND Direct2dViewer_getWindowHandler(void *D2dV)
 	return D2d->getWindowHandler();
 }
 
-void Direct2dViewer_setGammaValue( void *D2dV, UINT white, UINT black )
+void Direct2dViewer_setGammaValue( void *D2dV, UINT16 white, UINT16 black )
 {
 	Direct2dViewer *D2d = (Direct2dViewer *)D2dV;
 	D2d->SetGammaValue( white, black );
